requirements.txt
chainer==2.0.0b1
cupy==1.0.0b1
cycler==0.10.0
decorator==4.0.11
filelock==2.0.8
matplotlib==2.0.0
networkx==1.11
nose==1.3.7
numpy==1.12.1
olefile==0.44
Pillow==4.1.0
protobuf==3.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyWavelets==0.5.2
scikit-image==0.13.0
scipy==0.19.0
six==1.10.0
